Update create server page to server component and fetch auth there instead of on client
